日報登録、日報削除機能実装・テスト、APIテストデータreport.json作成、日報編集機能だけ修正の必要あり

+++ backend/src/main/java/com/daily_app/demo/Repository/DailyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.daily_app.demo.Entity.Daily;
 
 @Repository
-public interface DailyRepository extends JpaRepository<Daily, Long> {
+public interface DailyRepository extends JpaRepository<Daily, Integer> {
     
 }
